+++ Backend/src/main/java/com/ssafy/d108/backend/global/exception/ProductNotFoundException.java
@@ -11,7 +11,7 @@ public class ProductNotFoundException extends BusinessException {
         super(ErrorCode.PRODUCT_NOT_FOUND);
     }
 
-    public ProductNotFoundException(Long productId) {
+    public ProductNotFoundException(Integer productId) {
         super(ErrorCode.PRODUCT_NOT_FOUND, "상품을 찾을 수 없습니다. ID: " + productId);
     }
